feat: order create init

+++ mall-app/src/main/java/cn/eyeo/mall/app/order/inventory/InventoryApi.java
@@ -15,7 +15,7 @@ public interface InventoryApi {
      * @param count      购买数量
      * @return true-充足 false-库存不足
      */
-    boolean stockEnough(Long goodsSkuId, Long count);
+    boolean stockEnough(Long goodsSkuId, Integer count);
 
     /**
      * 锁定商品库存
@@ -24,7 +24,7 @@ public interface InventoryApi {
      * @param count      购买数量
      * @return true-锁定成功 false-锁定失败
      */
-    boolean lockStock(Long goodsSkuId, Long count);
+    boolean lockStock(Long goodsSkuId, Integer count);
 
     /**
      * 释放锁定的商品库存(Need Retry)
@@ -33,7 +33,7 @@ public interface InventoryApi {
      * @param count      购买数量
      * @return true-释放成功 false-释放失败
      */
-    boolean unlockStock(Long goodsSkuId, Long count);
+    boolean unlockStock(Long goodsSkuId, Integer count);
 
     /**
      * 扣减商品库存
@@ -42,7 +42,7 @@ public interface InventoryApi {
      * @param count      购买数量
      * @return true-扣减成功 false-扣减失败
      */
-    boolean reduceStock(Long goodsSkuId, Long count);
+    boolean reduceStock(Long goodsSkuId, Integer count);
 
     /**
      * 增加商品库存
@@ -51,6 +51,6 @@ public interface InventoryApi {
      * @param count      购买数量
      * @return true-增加成功 false-增加失败
      */
-    boolean addStock(Long goodsSkuId, Long count);
+    boolean addStock(Long goodsSkuId, Integer count);
 
 }
