+++ mall-app/src/main/java/cn/eyeo/mall/app/order/inventory/InventoryApiImpl.java
@@ -16,28 +16,28 @@ public class InventoryApiImpl implements InventoryApi {
     private static final Random RANDOM = new Random();
 
     @Override
-    public boolean stockEnough(Long goodsSkuId, Long count) {
+    public boolean stockEnough(Long goodsSkuId, Integer count) {
         return RANDOM.nextBoolean();
     }
 
     @Override
-    public boolean lockStock(Long goodsSkuId, Long count) {
+    public boolean lockStock(Long goodsSkuId, Integer count) {
         return RANDOM.nextBoolean();
     }
 
     @Override
-    public boolean unlockStock(Long goodsSkuId, Long count) {
+    public boolean unlockStock(Long goodsSkuId, Integer count) {
         // need retry
         return true;
     }
 
     @Override
-    public boolean reduceStock(Long goodsSkuId, Long count) {
+    public boolean reduceStock(Long goodsSkuId, Integer count) {
         return RANDOM.nextBoolean();
     }
 
     @Override
-    public boolean addStock(Long goodsSkuId, Long count) {
+    public boolean addStock(Long goodsSkuId, Integer count) {
         // need retry
         return true;
     }
